Use ComputerListener since it does not require builds to generate the metrics
